keep files by default in docs root unless they are different in older version, specify menus as much as possible in markdown

+++ generate.sh
@@ -2,7 +2,7 @@
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR
-git checkout docs/generated
+git checkout docs/*/generated
 rm -rf docs/*/cache
 mkdir docs/*/cache
 rm -rf docs/*/generated
@@ -19,7 +19,14 @@ git submodule update --recursive --force
 php composer.phar install || true
 cd ..
 php generator/generate.php --branch=master --targetname=2.x
+GENERATION_SUCCESS=$?
+
+if [ $GENERATION_SUCCESS -ne 0 ]; then
+  exit 1;
+fi
+
 php generator/generate.php --branch=3.x-dev --targetname=3.x
+
 GENERATION_SUCCESS=$?
 rm -rf app/tmp/cache/*
